+++ packages/babel-plugin-auto-logger/__e2e__/fixtures/custom-options/logger-output-object-args-as-object/input.js
@@ -40,6 +40,12 @@ function multiply(a, b) {
     const catchBlock = 'demo purpose for multiply';
   }
 
+  try {
+    const text = 'simple catch';
+  } catch {
+    const catchBlock = 'demo purpose for multiply';
+  }
+
   return a * b;
 }